refactor: Remove cls render prop argument
BREAKING CHANGE
The cls property was not adding much, since we are not using it
as a hook for styling in the library. There is an argument that
we could use it for fallbacks, but that requires the user to
set the style tag anyway. Thus, it is best to leave it undefined.
It also bloated the syntax quite a bit, with a mess of squgglies
that can easily be messed up; JSX prop, arrow function, and then
destructuring. {({cls}) => ... }. I think this is more convenient
for people to use, especially weighed against the lack of benefit
to the library.