Feature: Debounced/delayed Loading (#11)
* Add utils and delayed loading test in stories

* Intermediate solution

Not sure where the reducer/update function is headed, but we'll
find out, I suppose.

* Finish refactoring

* Make buffering work

It works, but now the intersection logic is not the same. Scrolling
until inView is true, and then up, triggers the promise cancelation.
This is undesirable. We want the image to keep loading as long as
it stays on the screen. Oddly, scrolling past the image, and it
being in view, still lets it load without canceling (as intendeed).
Another difference is that loading only commences once the image is
fully in view, rather than on the margin.

* Document things a bit better

* There was no bug; the placeholder had a 0 height!

* A few more docs

* Name things, expose as props

Added:
- debugActions: logs actions in dev, warns and logs in prod
- debounceDurationMs: whether, and how much to debounce loading,
  0 by default, so no change to v1. Added codepath to make it so.
- Provisional changes.md document with API compatibility
- Various TODOs and cleanup tasks

* Fix passing of debounceDurationMs to cb

* Use debugActions in the story, destructure it

Also add a note to hopefully remember to destructure next time :)

* Pin microbundle and ts versions

* Clean dir before build

* TS BUG: need declaration: false to build correctly

If we set declaration: true, then typescript (specifically rpt2,
Rollup's TS plugin) fails with a cryptic path error. This is
filed under https://github.com/Microsoft/TypeScript/issues/25047.

Curiously, this does not happen on master, and I have vaguely
traced it back to unionize and the fact that it does fancy type
exports, which we partially re-export here. We would have to
manually emit a declaration file, which is not hard tbh, but a bit
annoying. Even funnier, webpack has no issue building in storybook,
but that might be because it does not emit declarations anyway.

Steps:
- Get declaration file from master, add to repository, and copy to
dist/ upon build;
- See if massaging our exports fixes it;
- Wait for TS :/

* 1.1.0-rc.1

* Version bump

* Make update static, use Commands

The Elm/JS/Redux/React-Component fusion is complete. The reducer
is now a pure function, because it only describes side-effects,
and does not execute them. The logic is really nice to read there.

* Naming things

* Version bump

* Use TS3.0

* Copy definitions separately

Because of the "path undefined" bug, we have to run `tsc`
separately, to get the definitions and copy them over.

* Do not expose Buffering as a separate state

Expose it under 'Loading' for backwards compatibility

* Improve debounceDurationMs docs

* Add first point to README

* Cancel Promises, clear cache on unmount

* Add docs for debounceDurationMs